beater/config: split InstrumentationConfig out
